backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
pgvector==0.3.6
alembic==1.14.1

# Async processing
celery[redis]==5.4.0
redis==5.2.1

# AI / ML
openai==1.58.1
google-generativeai==0.8.5
numpy==2.2.1
scikit-learn==1.6.1
hdbscan==0.8.40

# Data collection
httpx==0.28.1
feedparser==6.0.11
beautifulsoup4==4.12.3
youtube-transcript-api==0.6.3

# Utilities
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
python-dateutil==2.9.0
langdetect==1.0.9
